requirements.txt
# ---- app core ----
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai>=1.51.0
pydantic>=2.7.0
pydantic-settings>=2.0.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# http client (compatible with fastapi-cloud-cli >=0.27.0)
httpx>=0.27.0,<0.29

# ---- langchain / langgraph family ----
# keep versions mutually compatible and satisfy langgraph-prebuilt's requirement
langchain-core>=0.3.67,<0.4
langchain>=0.3.7,<0.4
langchain-community>=0.3.7,<0.4
langchain-openai>=0.2.7,<0.3
langgraph==0.2.34
# (Optional) only add if you actually use it; otherwise omit to reduce surface area:
# langgraph-prebuilt==1.0.1

# ---- url_builder deps ----
geopy==2.4.1
airports-py==2.0.0
certifi>=2024.8.30